made forth task and some fixes

+++ src/task3.cpp
@@ -2,7 +2,7 @@
 unsigned long long sumPrime(unsigned int hbound)
 {
 	unsigned long long sum = 0;
-	for (int i = 2; i < hbound; i++)
+	for (long long i = 2; i < hbound; i++)
 	{
 		if(checkPrime(i) == true)
 			sum = sum + i;
